+++ PGPatcher/src/ParallaxGenConfig.cpp
@@ -101,14 +101,10 @@ void ParallaxGenConfig::loadConfig()
 
             m_userConfig = j;
             loadedConfig = true;
-            Logger::info("Loaded user config successfully");
-        } else {
-            Logger::error("Failed to parse ParallaxGen config file");
         }
     }
 
     if (!loadedConfig) {
-        Logger::warn("No user config found, using defaults");
         m_params = getDefaultParams();
     }
 }
@@ -170,6 +166,12 @@ auto ParallaxGenConfig::addConfigJSON(const nlohmann::json& j) -> void
         if (paramJ.contains("processing") && paramJ["processing"].contains("pluginlang")) {
             m_params.Processing.pluginLang
                 = ParallaxGenPlugin::getPluginLangFromString(paramJ["processing"]["pluginlang"].get<string>());
+        }
+        if (paramJ.contains("processing") && paramJ["processing"].contains("enabledebuglogging")) {
+            paramJ["processing"]["enabledebuglogging"].get_to<bool>(m_params.Processing.enableDebugLogging);
+        }
+        if (paramJ.contains("processing") && paramJ["processing"].contains("enabletracelogging")) {
+            paramJ["processing"]["enabletracelogging"].get_to<bool>(m_params.Processing.enableTraceLogging);
         }
 
         // "prepatcher"
@@ -436,6 +438,8 @@ auto ParallaxGenConfig::getUserConfigJSON() const -> nlohmann::json
     j["params"]["processing"]["pluginesmify"] = m_params.Processing.pluginESMify;
     j["params"]["processing"]["pluginlang"]
         = ParallaxGenPlugin::getStringFromPluginLang(m_params.Processing.pluginLang);
+    j["params"]["processing"]["enabledebuglogging"] = m_params.Processing.enableDebugLogging;
+    j["params"]["processing"]["enabletracelogging"] = m_params.Processing.enableTraceLogging;
 
     // "prepatcher"
     j["params"]["prepatcher"]["disablemlp"] = m_params.PrePatcher.disableMLP;